face/requirement.txt
amqp==5.0.5
anyjson==0.3.3
asgiref==3.3.1
autopep8==1.5.5
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.10
django-admin-generator==2.2.1
django-allauth==0.44.0
django-crispy-forms==1.11.0
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.1
django-utils2==2.11.3
djangorestframework==3.12.2
docx==0.2.4
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.4.0
jdcal==1.4.1
kombu==5.0.2
lxml==4.6.2
MouseInfo==0.1.3
mysql-connector-python==8.0.23
numpy==1.20.1
oauthlib==3.1.0
opencv-contrib-python==4.5.1.48
opencv-python==3.4.10.35
openpyxl==3.0.6
pandas==1.2.2
Pillow==8.1.0
prompt-toolkit==3.0.16
protobuf==3.15.3
psycopg2==2.8.6
PyAutoGUI==0.9.52
pycodestyle==2.6.0
pycparser==2.20
pydotplus==2.0.2
PyGetWindow==0.0.9
PyJWT==2.0.1
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.7
python-dateutil==2.8.1
python-decouple==3.4
python-docx==0.8.10
python-utils==2.5.6
python3-openid==3.2.0
pytube==10.4.1
PyTweening==1.0.3
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
Unipath==1.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.2.0
windows-curses==2.2.0
zipp==3.4.0
